requirements.txt
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
docopt==0.6.2
fonttools==4.34.4
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pip==22.1.2
pipreqs==0.4.11
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
setuptools==61.2.0
six==1.16.0
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.10
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
yarg==0.1.9
